Rename plugins package and move its confg
