+++ pipelex/cogt/llm/llm_worker_factory.py
@@ -22,13 +22,13 @@ class LLMWorkerFactory:
         llm_worker: LLMWorkerAbstract
         match llm_engine.llm_platform:
             case LLMPlatform.OPENAI | LLMPlatform.AZURE_OPENAI | LLMPlatform.PERPLEXITY | LLMPlatform.XAI:
-                from pipelex.plugin.openai.openai_factory import OpenAIFactory
+                from pipelex.plugins.openai.openai_factory import OpenAIFactory
 
                 structure_method: Optional[StructureMethod] = None
                 if get_config().cogt.llm_config.instructor_config.is_openai_structured_output_enabled:
                     structure_method = StructureMethod.INSTRUCTOR_OPENAI_STRUCTURED
 
-                from pipelex.plugin.openai.openai_llm_worker import OpenAILLMWorker
+                from pipelex.plugins.openai.openai_llm_worker import OpenAILLMWorker
 
                 llm_sdk_instance = plugin_manager.get_llm_sdk_instance(llm_sdk_handle=llm_sdk_handle) or plugin_manager.set_llm_sdk_instance(
                     llm_sdk_handle=llm_sdk_handle,
@@ -47,8 +47,8 @@ class LLMWorkerFactory:
                 except ImportError as exc:
                     raise MissingDependencyError("google-auth-oauthlib", "google", "This dependency is required to connect to google.") from exc
 
-                from pipelex.plugin.openai.openai_factory import OpenAIFactory
-                from pipelex.plugin.openai.openai_llm_worker import OpenAILLMWorker
+                from pipelex.plugins.openai.openai_factory import OpenAIFactory
+                from pipelex.plugins.openai.openai_llm_worker import OpenAILLMWorker
 
                 llm_sdk_instance = plugin_manager.get_llm_sdk_instance(llm_sdk_handle=llm_sdk_handle) or plugin_manager.set_llm_sdk_instance(
                     llm_sdk_handle=llm_sdk_handle,
@@ -62,8 +62,8 @@ class LLMWorkerFactory:
                     report_delegate=report_delegate,
                 )
             case LLMPlatform.CUSTOM_OPENAI:
-                from pipelex.plugin.openai.openai_factory import OpenAIFactory
-                from pipelex.plugin.openai.openai_llm_worker import OpenAILLMWorker
+                from pipelex.plugins.openai.openai_factory import OpenAIFactory
+                from pipelex.plugins.openai.openai_llm_worker import OpenAILLMWorker
 
                 llm_sdk_instance = plugin_manager.get_llm_sdk_instance(llm_sdk_handle=llm_sdk_handle) or plugin_manager.set_llm_sdk_instance(
                     llm_sdk_handle=llm_sdk_handle,
@@ -88,8 +88,8 @@ class LLMWorkerFactory:
                         (eg: bedrock-anthropic-claude)",
                     ) from exc
 
-                from pipelex.plugin.anthropic.anthropic_factory import AnthropicFactory
-                from pipelex.plugin.anthropic.anthropic_llm_worker import AnthropicLLMWorker
+                from pipelex.plugins.anthropic.anthropic_factory import AnthropicFactory
+                from pipelex.plugins.anthropic.anthropic_llm_worker import AnthropicLLMWorker
 
                 llm_sdk_instance = plugin_manager.get_llm_sdk_instance(llm_sdk_handle=llm_sdk_handle) or plugin_manager.set_llm_sdk_instance(
                     llm_sdk_handle=llm_sdk_handle,
@@ -114,8 +114,8 @@ class LLMWorkerFactory:
                         (eg: bedrock-mistral-large)",
                     ) from exc
 
-                from pipelex.plugin.mistral.mistral_factory import MistralFactory
-                from pipelex.plugin.mistral.mistral_llm_worker import MistralLLMWorker
+                from pipelex.plugins.mistral.mistral_factory import MistralFactory
+                from pipelex.plugins.mistral.mistral_llm_worker import MistralLLMWorker
 
                 llm_sdk_instance = plugin_manager.get_llm_sdk_instance(llm_sdk_handle=llm_sdk_handle) or plugin_manager.set_llm_sdk_instance(
                     llm_sdk_handle=llm_sdk_handle,
@@ -137,8 +137,8 @@ class LLMWorkerFactory:
                         "boto3,aioboto3", "bedrock", "The boto3 and aioboto3 SDKs are required to use Bedrock models."
                     ) from exc
 
-                from pipelex.plugin.bedrock.bedrock_factory import BedrockFactory
-                from pipelex.plugin.bedrock.bedrock_llm_worker import BedrockLLMWorker
+                from pipelex.plugins.bedrock.bedrock_factory import BedrockFactory
+                from pipelex.plugins.bedrock.bedrock_llm_worker import BedrockLLMWorker
 
                 llm_sdk_instance = plugin_manager.get_llm_sdk_instance(llm_sdk_handle=llm_sdk_handle) or plugin_manager.set_llm_sdk_instance(
                     llm_sdk_handle=llm_sdk_handle,
