+++ pipelex/config.py
@@ -8,6 +8,7 @@ from pipelex.exceptions import PipelexError
 from pipelex.hub import get_required_config
 from pipelex.libraries.library_config import LibraryConfig
 from pipelex.mission.track.tracker_config import TrackerConfig
+from pipelex.plugins.plugins_config import PluginsConfig
 from pipelex.tools.aws.aws_config import AwsConfig
 from pipelex.tools.config.models import ConfigModel, ConfigRoot
 from pipelex.tools.log.log_config import LogConfig
@@ -65,6 +66,7 @@ class Pipelex(ConfigModel):
 class PipelexConfig(ConfigRoot):
     session_id: str = shortuuid.uuid()
     cogt: Cogt
+    plugins: PluginsConfig
     pipelex: Pipelex
