+++ pipelex/plugins/anthropic/anthropic_llm_worker.py
@@ -14,7 +14,7 @@ from pipelex.cogt.llm.llm_models.llm_platform import LLMPlatform
 from pipelex.cogt.llm.llm_worker_abstract import LLMWorkerAbstract
 from pipelex.cogt.llm.structured_output import StructureMethod
 from pipelex.config import get_config
-from pipelex.plugin.anthropic.anthropic_factory import AnthropicFactory
+from pipelex.plugins.anthropic.anthropic_factory import AnthropicFactory
 from pipelex.tools.typing.pydantic_utils import BaseModelTypeVar
 
 
@@ -64,7 +64,7 @@ class AnthropicLLMWorker(LLMWorkerAbstract):
     # TODO: implement streaming behind the scenes to avoid timeout/streaming errors with Claude 4 and high tokens
     def _adapt_max_tokens(self, max_tokens: Optional[int]) -> int:
         max_tokens = max_tokens or self.default_max_tokens
-        if claude_4_tokens_limit := get_config().cogt.llm_config.anthropic_config.claude_4_tokens_limit:
+        if claude_4_tokens_limit := get_config().plugins.anthropic_config.claude_4_tokens_limit:
             if max_tokens > claude_4_tokens_limit:
                 max_tokens = claude_4_tokens_limit
                 log.warning(f"Max tokens is greater than the claude 4 reduced tokens limit, reducing to {max_tokens}")
