+++ pipelex/plugins/mistral/mistral_llm_worker.py
@@ -13,7 +13,7 @@ from pipelex.cogt.llm.llm_job_func import llm_job_func
 from pipelex.cogt.llm.llm_models.llm_engine import LLMEngine
 from pipelex.cogt.llm.llm_worker_abstract import LLMWorkerAbstract
 from pipelex.cogt.llm.structured_output import StructureMethod
-from pipelex.plugin.mistral.mistral_factory import MistralFactory
+from pipelex.plugins.mistral.mistral_factory import MistralFactory
 from pipelex.tools.typing.pydantic_utils import BaseModelTypeVar
