+++ pipelex/plugins/openai/openai_factory.py
@@ -31,7 +31,7 @@ class OpenAIFactory:
         api_key: Optional[str] = None
         match llm_platform:
             case LLMPlatform.AZURE_OPENAI:
-                azure_openai_config = get_config().cogt.llm_config.azure_openai_config
+                azure_openai_config = get_config().plugins.azure_openai_config
                 endpoint, api_version, api_key = azure_openai_config.configure(secrets_provider=get_secrets_provider())
 
                 log.verbose(f"Making AsyncAzureOpenAI client with endpoint: {endpoint}, api_version: {api_version}")
@@ -41,7 +41,7 @@ class OpenAIFactory:
                     api_version=api_version,
                 )
             case LLMPlatform.PERPLEXITY:
-                perplexity_config = get_config().cogt.llm_config.perplexity_config
+                perplexity_config = get_config().plugins.perplexity_config
                 endpoint, api_key = perplexity_config.configure(secrets_provider=get_secrets_provider())
 
                 log.verbose(f"Making perplexity AsyncOpenAI client with endpoint: {endpoint}")
@@ -50,11 +50,11 @@ class OpenAIFactory:
                     base_url=endpoint,
                 )
             case LLMPlatform.OPENAI:
-                openai_openai_config = get_config().cogt.llm_config.openai_openai_config
+                openai_openai_config = get_config().plugins.openai_config
                 api_key = openai_openai_config.get_api_key(secrets_provider=get_secrets_provider())
                 the_client = openai.AsyncOpenAI(api_key=api_key)
             case LLMPlatform.VERTEXAI_OPENAI:
-                vertexai_config = get_config().cogt.llm_config.vertexai_config
+                vertexai_config = get_config().plugins.vertexai_config
                 endpoint, api_key = vertexai_config.configure(secrets_provider=get_secrets_provider())
 
                 log.verbose(f"Making vertex AsyncOpenAI client with endpoint: {endpoint}")
@@ -63,7 +63,7 @@ class OpenAIFactory:
                     base_url=endpoint,
                 )
             case LLMPlatform.XAI:
-                xai_config = get_config().cogt.llm_config.xai_config
+                xai_config = get_config().plugins.xai_config
                 endpoint, api_key = xai_config.configure(secrets_provider=get_secrets_provider())
 
                 log.verbose(f"Making Xai AsyncOpenAI client with endpoint: {endpoint}")
@@ -72,7 +72,7 @@ class OpenAIFactory:
                     base_url=endpoint,
                 )
             case LLMPlatform.CUSTOM_OPENAI:
-                custom_endpoint_config = get_config().cogt.llm_config.custom_endpoint_config
+                custom_endpoint_config = get_config().plugins.custom_endpoint_config
                 base_url, api_key = custom_endpoint_config.configure(secrets_provider=get_secrets_provider())
 
                 log.verbose(f"Making custom AsyncOpenAI client with base_url: {base_url}")
