+++ pipelex/plugins/plugins_config.py
@@ -1,0 +1,30 @@
+from typing import Dict, List, Optional, cast
+
+from pydantic import Field, field_validator
+
+from pipelex.cogt.imgg.imgg_handle import ImggHandle
+from pipelex.cogt.imgg.imgg_job_components import ImggJobConfig, ImggJobParams, ImggJobParamsDefaults
+from pipelex.cogt.llm.llm_job_components import LLMJobConfig
+from pipelex.cogt.llm.llm_models.llm_platform import LLMPlatform
+from pipelex.plugins.anthropic.anthropic_config import AnthropicConfig
+from pipelex.plugins.bedrock.bedrock_config import BedrockConfig
+from pipelex.plugins.mistral.mistral_config import MistralConfig
+from pipelex.plugins.openai.azure_openai_config import AzureOpenAIConfig
+from pipelex.plugins.openai.custom_endpoint_config import CustomEndpointConfig
+from pipelex.plugins.openai.openai_config import OpenAIConfig
+from pipelex.plugins.openai.perplexity_config import PerplexityConfig
+from pipelex.plugins.openai.vertexai_config import VertexAIConfig
+from pipelex.plugins.openai.xai_config import XaiConfig
+from pipelex.tools.config.models import ConfigModel
+
+
+class PluginsConfig(ConfigModel):
+    anthropic_config: AnthropicConfig
+    azure_openai_config: AzureOpenAIConfig
+    bedrock_config: BedrockConfig
+    vertexai_config: VertexAIConfig
+    mistral_config: MistralConfig
+    openai_config: OpenAIConfig
+    perplexity_config: PerplexityConfig
+    xai_config: XaiConfig
+    custom_endpoint_config: CustomEndpointConfig
