+++ tests/pipelex/plugins/test_anthropic.py
@@ -5,7 +5,7 @@ from rich.table import Table
 
 from pipelex.cogt.exceptions import LLMSDKError
 from pipelex.cogt.llm.llm_models.llm_platform import LLMPlatform
-from pipelex.plugin.anthropic.anthropic_llms import anthropic_list_anthropic_models
+from pipelex.plugins.anthropic.anthropic_llms import anthropic_list_anthropic_models
 
 
 # make t VERBOSE=2 TEST=TestAnthropic
