+++ tests/pipelex/plugins/test_bedrock.py
