+++ tests/pipelex/plugins/test_openai.py
@@ -5,7 +5,7 @@ from pipelex.cogt.exceptions import LLMSDKError
 from pipelex.cogt.llm.llm_models.llm_platform import LLMPlatform
 from pipelex.config import get_config
 from pipelex.hub import get_secrets_provider
-from pipelex.plugin.openai.openai_llms import openai_list_available_models
+from pipelex.plugins.openai.openai_llms import openai_list_available_models
 
 
 # make t VERBOSE=2 TEST=TestOpenAI
@@ -14,7 +14,7 @@ from pipelex.plugin.openai.openai_llms import openai_list_available_models
 @pytest.mark.asyncio(loop_scope="class")
 class TestOpenAI:
     async def test_openai_api_key(self):
-        openai_openai_config = get_config().cogt.llm_config.openai_openai_config
+        openai_openai_config = get_config().plugins.openai_config
         assert openai_openai_config.get_api_key(secrets_provider=get_secrets_provider())
 
     # pytest -k test_openai_list_available_models -s -vv
